Started Boulder Class Implementation
Boulder can fall when all the dirt underneath it is dug out. However,
still a lot of bugs: the boulder has dirt behind it when it is created,
FrackMan can dig into the dirt behind the boulder (possibly caused by
the dirt behind the boulder?), haven’t checked to see if there is
another boulder beneath it when it is falling, nor implemented what it
would do when it hits a protestor, etc.